Scope proto definitions to package directory
Keeping these protobuf definitions in a flat file hierarchy as they are
kept upstream means there is no scoping for file locations for dependent
projects. That is, they must import files as `import "file.proto";`
rather than `import "android/bundle/file.proto";`. The former is not
clean and is likely to cause conflicts in downstream projects, while the
latter is much more preferable, although it requires us to slightly
modify the upstream files.